currentimage API (almost finished)
git-svn-id: http://svn.code.sf.net/p/rts-2/code/trunk/rts-2@10142 65ab8f4e-f147-0410-b3a9-f14133ecfe55

+++ lib/xmlrpc++/XmlRpcServerConnection.cpp
@@ -624,7 +624,7 @@ std::string XmlRpcServerConnection::generateHeader(std::string const& body)
 		"Content-length: ";
 
 	char buffLen[40];
-	sprintf(buffLen,"%i\r\n\r\n", (int) body.size());
+	sprintf(buffLen,"%zu\r\n\r\n", body.size());
 
 	return header + buffLen;
 }
